requirements.txt
flet==0.20.2
eyed3==0.9.7
pygame==2.5.2
GitPython==3.1.42